project1/ai-service/requirements.txt
langchain==0.2.0
fastapi==0.104.1
chromadb==0.4.15
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
uvicorn==0.24.0
requests==2.31.0
transformers==4.35.2
torch==2.1.0